% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_factor}
\alias{char_to_factor}
\title{Transmit levels of factor variable to parameter set}
\usage{
char_to_factor(levels)
}
\arguments{
\item{levels}{[\code{\link[=character]{character()}}] \cr Character vector of feature class labels.}
}
\description{
Transmit levels of factor variable to parameter set
}
