% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountARFactualClassif.R
\name{CountARFactualClassif}
\alias{CountARFactualClassif}
\title{CountARFactual (ARF-based Counterfactuals) for Classification Tasks}
\description{
CountARFactuals generates counterfactuals based on the adversarial random forests introduced in Watson et. al 2023
}
\examples{
if (require("randomForest")) {
 \donttest{
  # Train a model
  rf = randomForest(Species ~ ., data = iris)
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find counterfactuals for x_interest
  arf_classif = CountARFactualClassif$new(predictor)
 
  cfactuals = arf_classif$find_counterfactuals(
    x_interest = iris[150L, ], desired_class = "versicolor", desired_prob = c(0.5, 1)
  )
  # Print the counterfactuals
  cfactuals$data
  # Plot evolution of hypervolume and mean and minimum objective values
  cfactuals$evaluate_set()
  cfactuals$plot_parallel()
  }
}

}
\references{
Watson, D. S., Blesch, K., Kapar, J., and Wright, M. N. (2023). Adversarial
random forests for density estimation and generative modeling. In Proceedings
of the 26th International Conference on Artificial Intelligence and Statistics,
pages 5357–5375. PMLR.
}
\section{Super classes}{
\code{\link[counterfactuals:CounterfactualMethod]{counterfactuals::CounterfactualMethod}} -> \code{\link[counterfactuals:CounterfactualMethodClassif]{counterfactuals::CounterfactualMethodClassif}} -> \code{CountARFactualClassif}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CountARFactualClassif-new}{\code{CountARFactualClassif$new()}}
\item \href{#method-CountARFactualClassif-clone}{\code{CountARFactualClassif$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="counterfactuals" data-topic="CounterfactualMethod" data-id="print"><a href='../../counterfactuals/html/CounterfactualMethod.html#method-CounterfactualMethod-print'><code>counterfactuals::CounterfactualMethod$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="counterfactuals" data-topic="CounterfactualMethodClassif" data-id="find_counterfactuals"><a href='../../counterfactuals/html/CounterfactualMethodClassif.html#method-CounterfactualMethodClassif-find_counterfactuals'><code>counterfactuals::CounterfactualMethodClassif$find_counterfactuals()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountARFactualClassif-new"></a>}}
\if{latex}{\out{\hypertarget{method-CountARFactualClassif-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MOCClassif} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountARFactualClassif$new(
  predictor,
  max_feats_to_change = 1L,
  n_synth = 10L,
  n_iterations = 10L,
  feature_selector = "random_importance",
  importance_method = "fastshap"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with \code{iml::Predictor$new()}) holding the machine learning model and the data.}

\item{\code{max_feats_to_change}}{(\code{numeric(1)})\cr
The maximum number of features allowed to be altered (default 1L).}

\item{\code{n_synth}}{(\code{numeric(1)}) \cr
The number of samples drawn from the marginal distributions (default 10L).}

\item{\code{n_iterations}}{(\code{numeric(1)}) \cr
The number of iteration. In each iteration a new terminal node is chosen
from which the \code{n_synth} candidates are drawn.}

\item{\code{feature_selector}}{(\code{character(1)})\cr
The method to choose features that are fixed for a counterfactual and which
are part of the conditioning set when choosing tree paths. The default
\code{random_importance} means that the probability of being in the conditioning set
is proportional to how unimportant a feature is.
The strategy \code{importance} means that all features are chosen except for
(or less than) the \code{max_feats_to_change} features that are most important.
The strategy \code{random} randomly chooses up to \code{max_feats_to_change}
features that are not part of the conditioning set. All others are part.}

\item{\code{importance_method}}{(\code{character(1)})\cr
The local importance method to variables for the conditioning set.
Ignored if \code{feature_selector = "random"}.
Either "fastshap" based on the \code{fastshap} package (default) or "icesd"
based on the standard deviation of the ICE curve is possible.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountARFactualClassif-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CountARFactualClassif-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountARFactualClassif$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
