% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterfactualMethodRegr.R
\name{CounterfactualMethodRegr}
\alias{CounterfactualMethodRegr}
\title{Base class for Counterfactual Explanation Methods for Regression Tasks}
\description{
Abstract base class for counterfactual explanation methods for regression task.

\code{CounterfactualMethodRegr} can only be initialized for regression tasks. Child classes inherit the (public)
\verb{$find_counterfactuals()} method, which calls a (private) \verb{$run()} method. This \verb{$run()} method should be implemented
by the child classes and return the found counterfactuals as \code{data.table} (preferably) or \code{data.frame}.
TODO: Link to vignette with how to add new methods
}
\section{Inheritance}{

Child classes: \link{MOCRegr}, \link{WhatIfRegr}
}

\section{Super class}{
\code{\link[counterfactuals:CounterfactualMethod]{counterfactuals::CounterfactualMethod}} -> \code{CounterfactualMethodRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CounterfactualMethodRegr$new()}}
\item \href{#method-find_counterfactuals}{\code{CounterfactualMethodRegr$find_counterfactuals()}}
\item \href{#method-clone}{\code{CounterfactualMethodRegr$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="counterfactuals" data-topic="CounterfactualMethod" data-id="print">}\href{../../counterfactuals/html/CounterfactualMethod.html#method-print}{\code{counterfactuals::CounterfactualMethod$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new CounterfactualMethodRegr object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CounterfactualMethodRegr$new(predictor, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with \code{iml::Predictor$new()}) holding the machine learning model and the data.}

\item{\code{lower}}{(\code{numeric()} | \code{NULL})\cr
Vector of minimum values for numeric features. If not \code{NULL}, it should be named with the corresponding feature names.
If \code{NULL} (default) \code{lower} is taken for each numeric feature as its minimum value in \code{predictor$data$X}.}

\item{\code{upper}}{(\code{numeric()} | \code{NULL})\cr
Vector of maximum values for numeric features. If not \code{NULL}, it should be named with the corresponding feature names.
If \code{NULL} (default) \code{upper} is taken for each numeric feature as its maximum value in \code{predictor$data$X}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_counterfactuals"></a>}}
\if{latex}{\out{\hypertarget{method-find_counterfactuals}{}}}
\subsection{Method \code{find_counterfactuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CounterfactualMethodRegr$find_counterfactuals(x_interest, desired_outcome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_interest}}{(\code{data.table(1)} | \code{data.frame(1)}) \cr
A single row with the observation of interest.}

\item{\code{desired_outcome}}{(\code{numeric(1)} | \code{numeric(2)}) \cr
The desired predicted outcome. It can be a numeric scalar or a vector with two numeric values that specify an
outcome range.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{Counterfactuals} object containing the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CounterfactualMethodRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
