% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counterfactuals.R
\name{Counterfactuals}
\alias{Counterfactuals}
\title{Counterfactuals}
\description{
Counterfactuals

Counterfactuals
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Counterfactuals$new()}}
\item \href{#method-evaluate}{\code{Counterfactuals$evaluate()}}
\item \href{#method-predict}{\code{Counterfactuals$predict()}}
\item \href{#method-plot_parallel}{\code{Counterfactuals$plot_parallel()}}
\item \href{#method-plot_freq_of_feature_changes}{\code{Counterfactuals$plot_freq_of_feature_changes()}}
\item \href{#method-get_freq_of_feature_changes}{\code{Counterfactuals$get_freq_of_feature_changes()}}
\item \href{#method-plot_surface}{\code{Counterfactuals$plot_surface()}}
\item \href{#method-clone}{\code{Counterfactuals$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$new(cfactuals, predictor, x_interest, param_set, desired)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$evaluate(
  measures = c("dist_x_interest", "dist_target", "nr_changed"),
  show_diff = FALSE,
  k = 1,
  weights = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_parallel"></a>}}
\if{latex}{\out{\hypertarget{method-plot_parallel}{}}}
\subsection{Method \code{plot_parallel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$plot_parallel(feature_names = NULL, row_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_freq_of_feature_changes"></a>}}
\if{latex}{\out{\hypertarget{method-plot_freq_of_feature_changes}{}}}
\subsection{Method \code{plot_freq_of_feature_changes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$plot_freq_of_feature_changes(subset_zero = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_freq_of_feature_changes"></a>}}
\if{latex}{\out{\hypertarget{method-get_freq_of_feature_changes}{}}}
\subsection{Method \code{get_freq_of_feature_changes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$get_freq_of_feature_changes(subset_zero = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_surface"></a>}}
\if{latex}{\out{\hypertarget{method-plot_surface}{}}}
\subsection{Method \code{plot_surface()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$plot_surface(feature_names, grid_size = 250L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counterfactuals$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
