% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_to_interval}
\alias{dist_to_interval}
\title{Computes the (absolute, pairwise) distance between the vector elements and an interval}
\usage{
dist_to_interval(x, interval)
}
\arguments{
\item{x}{(\code{numeric()})\cr
A numeric vector.}

\item{interval}{(\code{numeric(2)})\cr
An interval.}
}
\description{
Computes the (absolute, pairwise) distance between the vector elements and an interval
}
