% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_counterfactuals.R
\name{fitness_fun}
\alias{fitness_fun}
\title{Compute fitness values}
\usage{
fitness_fun(
  x,
  x.interest,
  target,
  predictor,
  train.data,
  range = NULL,
  track.infeas = TRUE,
  identical.strategy = FALSE,
  k = 1,
  weights = NULL
)
}
\value{
(matrix)
}
\description{
Returns a matrix of three rows, one for the distance of the prediction to
the desired target, one for the distance to x.interest and
one for the number of changed features.
}
\section{Arguments}{

\describe{
\item{x: }{(data.frame)\cr Data frame of candidates.}
\item{x.interest: }{(data.frame)\cr Data frame with one row as instance to
be explained. }
\item{target: }{numeric(1)|numeric(2) \cr Desired outcome either a single numeric or
a vector of two numerics in order to define an interval as desired outcome.}
\item{predictor: }{(Predictor)\cr
The object that holds the machine learning model and the data.}
\item{range: }{(numeric)\cr Vector of ranges for numeric features.
Must have same ordering as columns in x and x.interest.}
}
}

