% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{get_diff}
\alias{get_diff}
\title{Get difference between two feature vectors}
\usage{
get_diff(x, x.interest)
}
\value{
(data.frame)
}
\description{
Difference is 0 if two characters or factors are equal.
Otherwise the class of x is displayed.
}
\section{Arguments}{

\describe{
\item{x: }{(data.frame)\cr Data frame with one row.}
\item{x.interest: }{(data.frame)\cr Data frame with one row.}
}
}

