% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_moc.R
\name{get_ICE_sd}
\alias{get_ICE_sd}
\title{Calculate ice curve standard deviation for all features}
\usage{
get_ICE_sd(x_interest, predictor, param_set)
}
\value{
(numeric) Vector with standard deviations for each feature.
}
\description{
Calculate ice curve standard deviation for all features
}
\section{Arguments}{

\describe{
\item{x_interest: }{(data.frame)\cr Data point, for which ice curves
should be calculated.}
\item{predictor: }{(Predictor)\cr Object, that holds the prediction model and dataset.}
\item{param_set: }{(ParamSet)\cr Parameter set of features in dataset.}
}
}

