% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{get_ICE_var}
\alias{get_ICE_var}
\title{Calculate ice curve variance over all features}
\usage{
get_ICE_var(x.interest, predictor, param.set)
}
\value{
(numeric) Vector with standard deviations for each feature.
}
\description{
Calculate ice curve variance over all features
}
\section{Arguments}{

\describe{
\item{x.interest: }{(data.frame)\cr Data point, for which ice curves
should be calculated.}
\item{predictor: }{(Predictor)\cr Object, that holds the prediction model and dataset.}
\item{param.set: }{(ParamSet)\cr Parameter set of features in dataset.}
}
}

