% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gower_dist_ranges}
\alias{gower_dist_ranges}
\title{Extract ranges from a \code{ParamSet} for computing the gower distance.}
\usage{
gower_dist_ranges(param_set)
}
\arguments{
\item{param_set}{[\link[ParamHelpers:makeParamSet]{ParamHelpers::ParamSet}].}
}
\value{
[\code{\link[=numeric]{numeric()}}] Ranges of all features.
}
\description{
For non-discrete features, the range is set to the difference between its upper and lower bounds. For
discrete features, it is set to \code{NA} as required by \link[StatMatch:gower.dist]{StatMatch::gower.dist}.
}
