% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{make_paramlist}
\alias{make_paramlist}
\title{Extract information about each feature from input dataset}
\usage{
make_paramlist(data, lower = NULL, upper = NULL)
}
\value{
(list)
}
\description{
Results in a list as input for ParamHelpers::makeParamSet.
}
\section{Arguments}{

\describe{
\item{data:}{(data.frame)\cr Training data}
\item{lower: }{numeric\cr Vector of minimal values for numeric features.
If NULL lower is extracted from input data specified in field 'data' of
'predictor'.}
\item{upper: }{numeric\cr Vector of maximal values for numeric features.
If NULL upper is extracted from input data specified in field 'data' of
'predictor'.}
}
}

