% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOCClassif.R
\name{MOCClassif}
\alias{MOCClassif}
\title{MOC for Classification Tasks}
\description{
\code{MOC} (Dandl et. al 2020) solves a multi-objective optimization problem to find counterfactuals. The four objectives
to minimize are:
\enumerate{
\item {Distance between \code{x_interest} and \code{desired_prob}}
\item {Distance between \code{x_interest} and a counterfactual}
\item {Number of feature changes (using the $L_0$ norm)}
\item {(Weighted) average distance between a counterfactual and the \code{k} nearest observed data points}
}

For optimization it uses
the NSGA II algorithm (Deb et. al 2002) with mixed integer evolutionary strategies by Li et al. (2013).
}
\details{
Several population initialization strategies are available:
\enumerate{
\item {\code{random}: Sample from numerical feature ranges and discrete feature values from \code{predictor$data$X}.
Some features values are randomly reset to the values of \code{x_interest}.}
\item {\code{icecurve}: Sample from numerical feature ranges and discrete feature values from \code{predictor$data$X}
as in \code{random} strategy . The higher the ICE curve variance of a feature, the lower the probability that
values of this feature are reset to the values of \code{x_interest}.}
\item {\code{sd}: Sample from numerical feature ranges that are limited by the feature standard deviations extracted
from \code{predictor$data$X}. Some features values are randomly reset to the values of \code{x_interest}.}
}

The R package \code{miesmuschel} implements the mixed integer evolutionary strategies.

The Gower's dissimilarity measure proposed by Kaufman and Rousseeuw (1990) computes all required distances.
It is implemented by \link[StatMatch]{gower.dist}.
}
\examples{
if (require("randomForest")) {
  # Train a model
  rf = randomForest(Species ~ ., data = iris)
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find counterfactuals
  moc_classif = MOCClassif$new(predictor)
  cfactuals = moc_classif$find_counterfactuals(
    x_interest = iris[150L, ], desired_class = "versicolor", desired_prob = c(0.5, 1)
  )
  # Print the results
  cfactuals$data
}

}
\references{
Binder Martin (2021). miesmuschel: Mixed Integer Evolutionary Strategies. R
package version 0.0.0-9000. https://github.com/mlr-org/miesmuschel

Dandl, Susanne, Christoph Molnar, Martin Binder, and Bernd Bischl. 2020. “Multi-Objective Counterfactual Explanations.”
In Parallel Problem Solving from Nature – PPSN XVI, edited by Thomas Bäck, Mike Preuss, André Deutz, Hao Wang, Carola Doerr,
Michael Emmerich, and Heike Trautmann, 448–69. Cham: Springer International Publishing.

Deb, K., Pratap, A., Agarwal, S., & Meyarivan, T. A. M. T. (2002). A fast and elitist multiobjective genetic algorithm: NSGA-II.
IEEE transactions on evolutionary computation, 6(2), 182-197.

R. Li et al., "Mixed Integer Evolution Strategies for Parameter Optimization," in Evolutionary Computation, vol. 21, no. 1,
pp. 29-64, March 2013, doi: 10.1162/EVCO_a_00059.
}
\section{Super classes}{
\code{\link[counterfactuals:CounterfactualMethod]{counterfactuals::CounterfactualMethod}} -> \code{\link[counterfactuals:CounterfactualMethodClassif]{counterfactuals::CounterfactualMethodClassif}} -> \code{MOCClassif}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MOCClassif$new()}}
\item \href{#method-plot_statistics}{\code{MOCClassif$plot_statistics()}}
\item \href{#method-get_dominated_hv}{\code{MOCClassif$get_dominated_hv()}}
\item \href{#method-plot_search}{\code{MOCClassif$plot_search()}}
\item \href{#method-clone}{\code{MOCClassif$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="counterfactuals" data-topic="CounterfactualMethod" data-id="print">}\href{../../counterfactuals/html/CounterfactualMethod.html#method-print}{\code{counterfactuals::CounterfactualMethod$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="counterfactuals" data-topic="CounterfactualMethodClassif" data-id="find_counterfactuals">}\href{../../counterfactuals/html/CounterfactualMethodClassif.html#method-find_counterfactuals}{\code{counterfactuals::CounterfactualMethodClassif$find_counterfactuals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MOCClassif$new(
  predictor,
  epsilon = NULL,
  fixed_features = NULL,
  max_changed = NULL,
  mu = 20L,
  n_generations = 175L,
  p_rec = 0.57,
  p_rec_gen = 0.85,
  p_rec_use_orig = 0.88,
  p_mut = 0.79,
  p_mut_gen = 0.56,
  p_mut_use_orig = 0.32,
  k = 1L,
  weights = NULL,
  lower = NULL,
  upper = NULL,
  init_strategy = "random",
  use_conditional_mutator = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-plot_statistics}{}}}
\subsection{Method \code{plot_statistics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MOCClassif$plot_statistics(centered_obj = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dominated_hv"></a>}}
\if{latex}{\out{\hypertarget{method-get_dominated_hv}{}}}
\subsection{Method \code{get_dominated_hv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MOCClassif$get_dominated_hv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_search"></a>}}
\if{latex}{\out{\hypertarget{method-plot_search}{}}}
\subsection{Method \code{plot_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MOCClassif$plot_search(objectives = c("dist_target", "dist_x_interest"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MOCClassif$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
