% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_counterfactuals.R
\name{mutConDens}
\alias{mutConDens}
\title{Mutator that uses conditional distributions.}
\usage{
mutConDens(ind, X, pred, param.set, ...)
}
\value{
(numeric(1)/character(1))
}
\description{
Mutator that uses conditional distributions.
}
\section{Arguments}{

\describe{
\item{ind: }{(numeric(1)/character(1))\cr Named vector of feature position.}
\item{X: }{(data table)\cr Candidate of which features are mutated.}
\item{pred: }{(Predictor)\cr The object that holds the machine learning model and the data.}
\item{param.set: }{(ParamSet)\cr Parameter set extracted from observed data.}
}
}

