% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamSetMaker.R
\name{ParamSetMaker}
\alias{ParamSetMaker}
\title{ParamSetMaker R6 Class}
\description{
Extracts information about each feature from input dataset
}
\details{
Based on a data set and optionally specified lower and upper bounds for each feature, it creates a
\link[ParamHelpers:makeParamSet]{ParamSet}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParamSetMaker$new()}}
\item \href{#method-make_param_set}{\code{ParamSetMaker$make_param_set()}}
\item \href{#method-clone}{\code{ParamSetMaker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ParamSetMaker} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamSetMaker$new(data, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{[\code{\link[=data.frame]{data.frame()}}] \cr A data set.}

\item{\code{lower}}{[\code{\link[=numeric]{numeric()}}]\cr Named vector of minimum values for numeric features.
A subset of features can also be specified. If NULL the minimum values are extracted from \code{data}.}

\item{\code{upper}}{[\code{\link[=numeric]{numeric()}}]\cr Named vector of maximum values for numeric features.
A subset of features can also be specified. If NULL the maximum values are extracted from \code{data}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ParamSetMaker} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_param_set"></a>}}
\if{latex}{\out{\hypertarget{method-make_param_set}{}}}
\subsection{Method \code{make_param_set()}}{
Makes a \link[ParamHelpers:makeParamSet]{ParamSet}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamSetMaker$make_param_set()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \link[ParamHelpers:makeParamSet]{ParamSet}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamSetMaker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
