% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counterfactuals.R
\name{plot.Counterfactuals}
\alias{plot.Counterfactuals}
\title{Plot Counterfactuals}
\format{
\code{\link{R6Class}} object.
}
\usage{
\method{plot}{Counterfactuals}(
  object,
  labels = FALSE,
  decimal.points = 3,
  nr.solutions = NULL,
  nr.changed = NULL
)
}
\value{
ggplot2 plot object
}
\description{
\code{plot.Counterfactuals()} plots the Pareto front in 2-D of the found Counterfactuals.
The x-axis displays the distance of the prediction to the desired prediction.
The y-axis displays the distance of the counterfactual to the original datapoint x.interest.
The colouring displays how many features were changed.
}
\section{Arguments}{

\describe{
\item{labels:}{logical(1)\cr Whether labels with difference to feature values of
x.interest should be plotted. Default is FALSE.}
\item{decimal.points:}{integer(1)\cr Number of decimal places used. Default is 3.}
\item{nr.solutions:}{integer(1)\cr Number of solutions showed. Default NULL means,
all solutions are showed.}
\item{nr.changed:}{integer(1)\cr Plot only counterfactuals with certain number of
changed features. Default is NULL.}
}
}

\examples{
\dontrun{
if (require("randomForest")) {
data("Boston", package  = "MASS")
Boston = Boston
set.seed(1000)
rf =  randomForest(medv ~ ., data = Boston)
X = Boston[-which(names(Boston) == "medv")]
mod = Predictor$new(rf, data = X)

# Then we calculate Counterfactuals for the first instance
x.interest = X[1,]
mod$predict(x.interest)
target = 30
cf = Counterfactuals$new(mod, x.interest = x.interest, target = target,
  mu = 50, generations = 60)

# The results can be plotted
plot(cf)
plot(cf, labels = TRUE, nr.solutions = 10)
}
}
}
\seealso{
\link{Counterfactuals}
}
