% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{round_df}
\alias{round_df}
\title{Round numeric elements in data frame}
\usage{
round_df(df, digits = 3)
}
\value{
(data.frame) with rounded elements
}
\description{
Round numeric elements in data frame
}
\section{Arguments}{

\describe{
\item{df: }{(data.frame)\cr Data frame in which numeric elments will be rounded.}
\item{digits: }{integer(1)\cr Number of decimal places used. Default is 3.}
}
}

