% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{sdev_to_list}
\alias{sdev_to_list}
\title{Create a list with named vectors of standard deviation}
\usage{
sdev_to_list(sdev, param.set)
}
\value{
(list)
}
\description{
Will be grouped by the feature type extracted from param.set.
}
\section{Arguments}{

\describe{
\item{sdev: }{(numeric)\cr Vector of standard deviations.}
\item{param.set: }{(ParamSet)\cr Paramset to create elements of list,
one for digit features and one for integer features.}
}
}

