% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smallest_n_indices}
\alias{smallest_n_indices}
\title{Returns the indices of the n smallest elements in a vector}
\usage{
smallest_n_indices(x, n = 1L)
}
\arguments{
\item{x}{(\code{numeric()})\cr
A numeric vector.}

\item{n}{(\code{numeric(1)})\cr
A integer indicating how many elements should be returned}
}
\description{
Returns the indices of the n smallest elements in a vector
}
