% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_moc.R
\name{transform_to_orig}
\alias{transform_to_orig}
\title{Get difference between two feature vectors}
\usage{
transform_to_orig(
  x,
  x_interest,
  delete_use_orig = FALSE,
  fixed_features = NULL,
  max_changed = NULL
)

transform_to_orig(
  x,
  x_interest,
  delete_use_orig = FALSE,
  fixed_features = NULL,
  max_changed = NULL
)
}
\value{
(data.frame)
get_diff = function(x, x.interest) {
assertDataFrame(x, ncols = ncol(x.interest))
assertDataFrame(x.interest, nrows = 1)

diff = data.frame(matrix(data = NA, nrow = nrow(x), ncol = length(x.interest)))
names(diff) = names(x.interest)

x.interest = x.interest\link{rep(row.names(x.interest), nrow(x)), }

for (i in 1:ncol(x)) {
if (class(x\link{,i}) == "character"|
class(x\link{,i}) == "factor") {
p.val = as.character(x\link{,i})
x.val = as.character(x.interest\link{,i})
diff\link{,i} = ifelse(p.val != x.val, p.val, "0")
} else {
diff\link{,i} = x\link{,i} - x.interest\link{,i}
}
}
return(diff)
}

Transform features to x.interest

Replace features of x by value of x.interest
where use.orig is set to TRUE.

(list)

(list)
}
\description{
Difference is 0 if two characters or factors are equal.
Otherwise the class of x is displayed.

Replace features of x by value of x.interest
where use.orig is set to TRUE.
}
\section{Arguments}{

\describe{
\item{x: }{(data.frame)\cr Data frame with one row.}
\item{x.interest: }{(data.frame)\cr Data frame with one row.}
}


\describe{
\item{x: }{(list)\cr List of features, must have list element
use.origin.}
\item{x.interest: }{(data.frame)\cr Data frame with one row.}
\item{delete.use.orig: }{(logical(1))\cr Whether to
delete list element use.orig of x.}
\item{fixed.features: }{(character)\cr
Indicate fixed features by feature name.}
\item{max.changed: }{numeric(1)\cr Number indicating
how many feature are allowed to maximially differ from the original data point.}
}


\describe{
\item{x: }{(list)\cr List of features, must have list element
use.origin.}
\item{x.interest: }{(data.frame)\cr Data frame with one row.}
\item{delete.use.orig: }{(logical(1))\cr Whether to
delete list element use.orig of x.}
\item{fixed.features: }{(character)\cr
Indicate fixed features by feature name.}
\item{max.changed: }{numeric(1)\cr Number indicating
how many feature are allowed to maximially differ from the original data point.}
}
}

