% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{char_to_factor}
\alias{char_to_factor}
\title{Transmit levels of factor variable to parameter set}
\usage{
char_to_factor(levels)
}
\description{
Transmit levels of factor variable to parameter set
}
\section{Arguments}{

\describe{
\item{levels: }{(character)\cr Character vector of feature class labels.}
}
}

