% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{transform_to_orig}
\alias{transform_to_orig}
\title{Transform features to x.interest}
\usage{
transform_to_orig(
  x,
  x.interest,
  delete.use.orig = FALSE,
  fixed.features = NULL,
  max.changed = NULL
)
}
\value{
(list)
}
\description{
Replace features of x by value of x.interest
where use.orig is set to TRUE.
}
\section{Arguments}{

\describe{
\item{x: }{(list)\cr List of features, must have list element
use.origin.}
\item{x.interest: }{(data.frame)\cr Data frame with one row.}
\item{delete.use.orig: }{(logical(1))\cr Whether to
delete list element use.orig of x.}
\item{fixed.features: }{(character)\cr
Indicate fixed features by feature name.}
\item{max.changed: }{numeric(1)\cr Number indicating
how many feature are allowed to maximially differ from the original data point.}
}
}

